Update password reminder information
We've had feedback that current guidance is unclear how non-government users can get password reminders, so we need to address this. In addtion we need to make clear end-users should contact local IT helpdesk for support, and network admins should review our tech docs for support in the first instance.